!1986 feat(#I4F78M): add ScrollMode parameter on Table component * doc: 更新示例 * doc: 增加行虚拟滚动示例 * feat: 增加虚拟化技术 * feat: 增加 StartIndex 参数 * chore: 增加 is-ph 样式 * feat: 增加 StartIndex  参数 * refactor: 更改 TotalCount 数据类型为 int * feat: 增加虚拟滚动参数 * feat: 增加虚拟滚动枚举 * doc: 增加虚拟滚动菜单 * doc: 增加虚拟滚动菜单资源文件 * doc: 增加 Ip 菜单资源文件 * feat: 增加 IsVirtulaizer 参数 * feat: 增加 IP 地址解析功能

+++ src/BootstrapBlazor/Components/Table/QueryData.cs
@@ -20,7 +20,7 @@ namespace BootstrapBlazor.Components
         /// <summary>
         /// 获得/设置 数据集合总数
         /// </summary>
-        public long TotalCount { get; set; }
+        public int TotalCount { get; set; }
 
         /// <summary>
         /// 获得/设置 数据是否被过滤 默认为 false 未被过滤
